add tests for profile

+++ app/Http/Controllers/Admin/AnswerController.php
@@ -78,6 +78,9 @@ class AnswerController extends Controller
         // IF Parent Answer delete All Child Answers
         $answer = $this->answerRepository->model->find($id);
 
+        if(!$answer) {
+            return redirect()->back()->with('error','incorrect access');
+        }
         $answer->childAnswers()->delete();
 
         $answer->question()->delete();
